fix: updated signin route to handle when no user found

+++ backend/src/controllers/user/signin.js
@@ -8,36 +8,44 @@ const verifyCredential = async (req, res) => {
     const { email, password } = req.body;
     try {
         const user = await User.findOne({ email: email })
-        if (user.isVerified && bcrypt.compareSync(password, user.password)) {//if user is verified and has correct password
-            const refreshToken = generateToken('refreshToken', user._id, '2h');
-            await Session.deleteMany({ userId: user._id });//deleting old sessions
-            const session = new Session({ userId: user._id, refreshToken });//new session
-            session.save();
+        if (user) {
+            if (user.isVerified && bcrypt.compareSync(password, user.password)) {//if user is verified and has correct password
+                const refreshToken = generateToken('refreshToken', user._id, '2h');
+                await Session.deleteMany({ userId: user._id });//deleting old sessions
+                const session = new Session({ userId: user._id, refreshToken });//new session
+                session.save();
 
-            res.status(200).send({
-                success: true,
-                message: "successfully logged in",
-                accessToken: generateToken('accessToken', user._id, '5h'),
-                refreshToken
-            })
-        } else {
-            if (!bcrypt.compareSync(password, user.password)) {
-                res.status(401).send({
-                    success: false,
-                    message: "Unable to login: Incorrect Credentials",
-                });
-            } else if (!user.isVerified) {
-                res.status(403).send({
-                    success: false,
-                    message: "Unable to login: User not verified",
-                });
+                res.status(200).send({
+                    success: true,
+                    message: "successfully logged in",
+                    accessToken: generateToken('accessToken', user._id, '5h'),
+                    refreshToken
+                })
             } else {
-                res.status(400).send({
-                    success: false,
-                    message: "Unable to login",
-                });
+                if (!bcrypt.compareSync(password, user.password)) {
+                    res.status(401).send({
+                        success: false,
+                        message: "Unable to login: Incorrect Credentials",
+                    });
+                } else if (!user.isVerified) {
+                    res.status(403).send({
+                        success: false,
+                        message: "Unable to login: User not verified",
+                    });
+                } else {
+                    res.status(400).send({
+                        success: false,
+                        message: "Unable to login",
+                    });
+                }
             }
+        }else{
+            res.status(404).send({
+                success:false,
+                message:"User not Found"
+            })
         }
+
     } catch (error) {
         console.log('err :>> ', error);
         res.status(500).send("Unable to fetch user")
